refactor: Separate business logic into custom hooks
